created MSystem and MClass
Created a relationship between MSystem and MClass to get all the classes in the system.
Created a relationship between MClasses to get all the superclasses of a certain class.
Created a relationship between MClasses to get all the subclasses of a certain class.
Created a metric to compute the number of casts to a certain class.