Created MMethod
Created a relationship between MSystem and MClass to get all the classes with the number of descendants casts greater than 1.
Created a metric to compute the number of classes in the system.
Created a metric to compute the number of lines of code in the system.
Created a relationship between MClass and MMethod to get all the methods of a certain class.
Created a relationship between MClass and MMethod to get all the methods with casts to a certain class.
Created a relationship between MClass and MMethod to get all the methods with casts to the descendants of a certain class.
Created a metric to compute the number of descendants casts of a certain class.
Created a metric to get the containing class of a certain method.